requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
attrs==22.1.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
decorator==5.1.1
Django==3.2.9
django-appconf==1.0.5
django-cryptography==1.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
iniconfig==1.1.1
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
mysqlclient==2.1.0
openpyxl==3.0.9
packaging==21.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
promise==2.3
prompt-toolkit==3.0.29
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.14.1
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.0.1
pytz==2022.1
requests==2.27.1
retry==0.9.2
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==1.2.3
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
